pyproject.toml
[project]
name = "coreindex"
version = "1.1.0"
description = "A lightweight Bitcoin indexer optimized for light clients and Silent Payments."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bitarray>=3.8.0",
    "coincurve>=21.0.0",
    "httpx>=0.28.1",
    "polars>=1.36.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
